Successfully extracted the leed function.
