+++ cleedpy/cleed/test_cleed.c
@@ -246,306 +246,10 @@ FILE *res_stream;
   fprintf(STDCTR, "(LEED_TEMP): n_set = %d\n", n_set);
 #endif
 
-/*********************************************************************
- Energy Loop
-*********************************************************************/
 
-  for( energy = eng->ini; energy < eng->fin + E_TOLERANCE; energy += eng->stp)
-  {
-
-    pc_update(v_par, phs_shifts, energy);
-    n_beams_now = bm_select(&beams_now, beams_all, v_par, bulk->dmin);
-
-#ifdef CONTROL
-    fprintf(STDCTR, "(LEED_TEMP):\n\t => E = %.1f eV (%d beams used) <=\n\n",
-              v_par->eng_v*HART, n_beams_now);
-#endif
-
-
-/*********************************************************************
-  BULK:
-  Loop over beam sets
-
-  Create matrix R_bulk that will eventually contain the bulk
-  reflection matrix
-*********************************************************************/
-
-    R_bulk = matalloc(R_bulk, n_beams_now, n_beams_now, NUM_COMPLEX);
-    for(offset = 1, i_set = 0; i_set < n_set; i_set ++)
-    {
-      n_beams_set = bm_set(&beams_set, beams_now, i_set);
-
-/*********************************************************************
-    Loop over periodic bulk layers
-*********************************************************************/
-
-    /**********************************************************
-     Compute scattering matrices for bottom-most bulk layer:
-     - single Bravais layer or composite layer
-    **********************************************************/
-
-#ifdef CONTROL_FLOW
-      fprintf(STDCTR, "(LEED_TEMP periodic): bulk layer %d/%d, set %d/%d\n",
-                        0, bulk->nlayers - 1, i_set, n_set - 1);
-#endif
-
-      if( (bulk->layers + 0)->natoms == 1)
-      {
-        ms_bravl_nd( &Tpp, &Tmm, &Rpm, &Rmp,
-                     v_par, (bulk->layers + 0), beams_set);
-      }
-      else
-      {
-        ms_compl_nd( &Tpp, &Tmm, &Rpm, &Rmp,
-                     v_par, (bulk->layers + 0), beams_set);
-      }
-
-#ifdef CONTROL_X
-      fprintf(STDCTR, "(LEED_TEMP): after ms_bravl_nd: Tpp:");
-      matshow(Tpp);
-      fprintf(STDCTR, "(LEED_TEMP): after ms_bravl_nd: Tmm:");
-      matshow(Tmm);
-      fprintf(STDCTR, "(LEED_TEMP): after ms_bravl_nd: Rpm:");
-      matshow(Rpm);
-      fprintf(STDCTR, "(LEED_TEMP): after ms_bravl_nd: Rmp:");
-      matshow(Rmp);
-#endif
-
-    /**********************************************************
-      Loop over the other bulk layers
-    **********************************************************/
-
-      for(i_layer = 1;
-          ( (bulk->layers+i_layer)->periodic == 1) &&
-          (i_layer < bulk->nlayers);
-          i_layer ++)
-      {
-#ifdef CONTROL_FLOW
-        fprintf(STDCTR, "(LEED_TEMP periodic): bulk layer %d/%d, set %d/%d\n",
-                        i_layer, bulk->nlayers - 1, i_set, n_set - 1);
-#endif
-
-    /**************************************************************
-      Compute scattering matrices R/T_s for a single bulk layer
-       - single Bravais layer or composite layer
-    ***************************************************************/
-
-        if( (bulk->layers + i_layer)->natoms == 1)
-        {
-          ms_bravl_nd ( &Tpp_s, &Tmm_s, &Rpm_s, &Rmp_s,
-                        v_par, (bulk->layers + i_layer), beams_set);
-        }
-        else
-        {
-          ms_compl_nd( &Tpp_s, &Tmm_s, &Rpm_s, &Rmp_s,
-                       v_par, (bulk->layers + i_layer), beams_set);
-        }
-
-    /***************************************************************************
-       Add the single layer matrices to the rest by layer doubling
-       - inter layer vector is the vector between layers
-         (i_layer - 1) and (i_layer):
-         (bulk->layers + i_layer)->vec_from_last
-    ****************************************************************************/
-#ifdef CONTROL_FLOW
-        fprintf(STDCTR,
-                "(LEED_TEMP): before ld_2lay vec_from...(%.2f %.2f %.2f)\n",
-                       (bulk->layers + i_layer)->vec_from_last[1] * BOHR,
-                       (bulk->layers + i_layer)->vec_from_last[2] * BOHR,
-                       (bulk->layers + i_layer)->vec_from_last[3] * BOHR);
-#endif
-
-        ld_2lay( &Tpp,  &Tmm,  &Rpm,  &Rmp,
-                 Tpp,   Tmm,   Rpm,   Rmp,
-                 Tpp_s, Tmm_s, Rpm_s, Rmp_s,
-                 beams_set, (bulk->layers + i_layer)->vec_from_last);
-
-      } /* for i_layer (bulk) */
-
-   /*********************************************************************
-      Layer doubling for all periodic bulk layers until convergence is
-      reached:
-       - inter layer vector is (bulk->layers + 0)->vec_from_last
-   **********************************************************************/
-#ifdef CONTROL_FLOW
-      fprintf(STDCTR, "(LEED_TEMP): before ld_2n vec_from...(%.2f %.2f %.2f)\n",
-                      (bulk->layers + 0)->vec_from_last[1] * BOHR,
-                      (bulk->layers + 0)->vec_from_last[2] * BOHR,
-                      (bulk->layers + 0)->vec_from_last[3] * BOHR);
-#endif
-
-      Rpm = ld_2n( Rpm, Tpp, Tmm, Rpm, Rmp,
-                   beams_set, (bulk->layers + 0)->vec_from_last);
-
-   /*******************************************************************
-     Compute scattering matrices for top-most bulk layer if it is
-     not periodic.
-      - single Bravais layer or composite layer
-   **********************************************************************/
-
-      if( i_layer == bulk->nlayers - 1 )
-      {
-#ifdef CONTROL_FLOW
-        fprintf(STDCTR,
-                "(LEED_TEMP not periodic): bulk layer %d/%d, set %d/%d\n",
-                i_layer, bulk->nlayers - 1, i_set, n_set - 1);
-#endif
-
-        if( (bulk->layers + i_layer)->natoms == 1)
-        {
-          ms_bravl_nd( &Tpp_s, &Tmm_s, &Rpm_s, &Rmp_s,
-                    v_par, (bulk->layers + i_layer), beams_set);
-        }
-        else
-        {
-          ms_compl_nd( &Tpp_s, &Tmm_s, &Rpm_s, &Rmp_s,
-                       v_par, (bulk->layers + i_layer), beams_set);
-        }
-
-    /**************************************************************************
-       Add the single layer matrices of the top-most layer to the rest
-       by layer doubling:
-       - inter layer vector is the vector between layers
-         (i_layer - 1) and (i_layer):
-         (bulk->layers + i_layer)->vec_from_last
-    ***************************************************************************/
-
-        Rpm = ld_2lay_rpm(Rpm, Rpm, Tpp_s, Tmm_s, Rpm_s, Rmp_s,
-                          beams_set, (bulk->layers + i_layer)->vec_from_last);
-
-      }  /* if( i_layer == bulk->nlayers - 1 ) */
-
-   /*******************************************************
-     Insert reflection matrix for this beam set into R_bulk.
-   ********************************************************/
-
-      R_bulk = matins(R_bulk, Rpm, offset, offset);
-      offset += n_beams_set;
-
-   /*************************
-     Write cpu time to output
-   **************************/
-
-      sprintf(linebuffer,"(LEED_TEMP): bulk layers set %d, E = %.1f",
-              i_set, energy*HART);
-      cpu_time(STDCPU,linebuffer);
-    }  /* for i_set */
-
-/*********************************************************************
-  OVERLAYER
-  Loop over all overlayer layers
-*********************************************************************/
-
-    for(i_layer = 0; i_layer < over->nlayers; i_layer ++)
-    {
-#ifdef CONTROL_FLOW
-      fprintf(STDCTR, "(LEED_TEMP): overlayer %d/%d\n", i_layer, over->nlayers - 1);
-#endif
-   /***********************************************************
-     Calculate scattering matrices for a single overlayer layer
-      - only single Bravais layer
-   ************************************************************/
-
-      if( (over->layers + i_layer)->natoms == 1)
-      {
-        ms_bravl_nd( &Tpp_s, &Tmm_s, &Rpm_s, &Rmp_s,
-                     v_par, (over->layers + i_layer), beams_now);
-      }
-      else
-      {
-        ms_compl_nd( &Tpp_s, &Tmm_s, &Rpm_s, &Rmp_s,
-                     v_par, (over->layers + i_layer), beams_now);
-      }
-
-#ifdef CONTROL_X
- fprintf(STDCTR, "\n(LEED_TEMP):overlayer %d  ...\n",i_layer);
- fprintf(STDCTR, "\n(LEED_TEMP): Tpp:\n");
- matshowabs(Tpp_s);
- fprintf(STDCTR, "\n(LEED_TEMP): Tmm:\n");
- matshowabs(Tmm_s);
- fprintf(STDCTR, "\n(LEED_TEMP): Rpm:\n");
- matshowabs(Rpm_s);
- fprintf(STDCTR, "\n(LEED_TEMP): Rmp:\n");
- matshowabs(Rmp_s);
-#endif
-
-  /****************************************************************
-     Add the single layer matrices to the rest by layer doubling:
-     - if the current layer is the bottom-most (i_layer == 0),
-       the inter layer vector is calculated from the vectors between
-       top-most bulk layer and origin
-       ( (bulk->layers + nlayers)->vec_to_next )
-       and origin and bottom-most overlayer
-       (over->layers + 0)->vec_from_last.
-
-     - inter layer vector is the vector between layers
-       (i_layer - 1) and (i_layer): (over->layers + i_layer)->vec_from_last
-  **********************************************************************/
-
-      if (i_layer == 0)
-      {
-        for(i_c = 1; i_c <= 3; i_c ++)
-        {
-          vec[i_c] = (bulk->layers + bulk->nlayers - 1)->vec_to_next[i_c]
-                     + (over->layers + 0)->vec_from_last[i_c];
-        }
-
-#ifdef CONTROL_FLOW
-        fprintf(STDCTR,
-                "(LEED):over0 before ld_2lay_rpm vec..(%.2f %.2f %.2f)\n",
-                vec[1] * BOHR,vec[2] * BOHR, vec[3] * BOHR);
-#endif
-
-        R_tot = ld_2lay_rpm(R_tot, R_bulk, Tpp_s, Tmm_s, Rpm_s, Rmp_s,
-                            beams_now, vec);
-      }
-      else
-      {
-#ifdef CONTROL_FLOW
-        fprintf(STDCTR,
-                "(LEED):over%d  before ld_2lay_rpm vec..(%.2f %.2f %.2f)\n",
-                i_layer,(over->layers + i_layer)->vec_from_last[1] * BOHR,
-                        (over->layers + i_layer)->vec_from_last[2] * BOHR,
-                        (over->layers + i_layer)->vec_from_last[3] * BOHR);
-#endif
-
-        R_tot = ld_2lay_rpm(R_tot, R_tot, Tpp_s, Tmm_s, Rpm_s, Rmp_s,
-                          beams_now, (over->layers + i_layer)->vec_from_last);
-      }
-
-   /**************************
-     Write cpu time to output
-   **************************/
-
-      sprintf(linebuffer,"(LEED): overlayer %d, E = %.1f",
-              i_layer, energy * HART);
-      cpu_time(STDCPU,linebuffer);
-
-    }  /* for i_layer (overlayer) */
-
-/*********************************************
-   Add propagation towards the potential step.
-**********************************************/
-
-    vec[1] = vec[2] = 0.;
-    vec[3] = 1.25 / BOHR;
-
-/********************************************
-    No scattering at pot. step
-********************************************/
-
-    Amp = ld_potstep0(Amp, R_tot, beams_now, v_par->eng_v, vec);
-    out_int(Amp, beams_now, beams_out, v_par, res_stream);
-
-/********************************************
-    Write cpu time to output
-********************************************/
-
-    sprintf(linebuffer,"  %.1f   %d  ",energy * HART,n_beams_now);
-    cpu_time(STDWAR,linebuffer);
-
-
-  } /* end of energy loop */
+  printf("Starting leed\n");
+  leed(bulk, over, phs_shifts, beams_all, beams_out, n_set, eng, v_par, res_stream);
+  printf("Finished leed\n");
 
 
 #ifdef CONTROL_IO
