+++ examples/cleed_example_old/run.sh
@@ -1,4 +1,4 @@
 export CLEED_PHASE=./PHASE
-EXECUTABLE=../../cleedpy/cleed/test_nsym_lp
+EXECUTABLE=../../cleedpy/cleed/build/test_cleed
 
 ${EXECUTABLE} -i Ni111_Cu.inp
